CAPS-83 New Portal functionality with mirrors
Alter the portal functionality to allow multiple "pairs" of portals to be created.
If a mirror is locked in place and the player creates a portal at the mirror it will create a portal to allow the beam of light to pass.
The portal will resize and match with the gameobject.
Once a successful mirror "pair" has been created the portal will disappear and activate particle effect